server/requirements.txt
flask
flask_cors
datetime
uuid
pyjwt==2.0.0